feat: use router and other fixes (#80)
